--- software/contrib/consequencer_patterns.py
@@ -1,172 +1,0 @@
-class pattern:
-
-	# Initialize pattern lists
-    BD=[]
-    SN=[]
-    HH=[]
-
-    # Add patterns
-    BD.append("1000100010001000")
-    SN.append("0000000000000000")
-    HH.append("0000000000000000")
-
-    BD.append("1000100010001000")
-    SN.append("0000000000000000")
-    HH.append("0010010010010010")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000000000")
-    HH.append("0010010010010010")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000001000")
-    HH.append("0010010010010010")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000000000")
-    HH.append("0000000000000000")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000001000")
-    HH.append("0000000000000000")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000001000")
-    HH.append("0000100010001001")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000001000")
-    HH.append("0101010101010101")
-
-    BD.append("1000100010001000")
-    SN.append("0000000000000000")
-    HH.append("1111111111111111")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000001000")
-    HH.append("1111111111111111")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000000000")
-    HH.append("0001000000000000")
-
-    BD.append("1000100010001000")
-    SN.append("0000100000000000")
-    HH.append("0001001000000000")
-
-    # Source: https://docs.google.com/spreadsheets/d/19_3BxUMy3uy1Gb0V8Wc-TcG7q16Amfn6e8QVw4-HuD0/edit#gid=0
-    BD.append("1000000010000000")
-    SN.append("0000100000001000")
-    HH.append("1010101010101010")
-
-    BD.append("1010001000100100")
-    SN.append("0000100101011001")
-    HH.append("0000000100000100")
-
-    BD.append("1000000110000010")
-    SN.append("0000100000001000")
-    HH.append("1010101110001010")
-
-    BD.append("1100000100110000")
-    SN.append("0000100000001000")
-    HH.append("1010101010101010")
-
-    BD.append("1000000110100000")
-    SN.append("0000100000001000")
-    HH.append("0010101010101010")
-
-    BD.append("1010000000110001")
-    SN.append("0000100000001000")
-    HH.append("1010101010101010")
-
-    BD.append("1000000110100001")
-    SN.append("0000100000001000")
-    HH.append("0000100010101011")
-
-    BD.append("1001001010000000")
-    SN.append("0000100000001000")
-    HH.append("0000100000001000")
-
-    BD.append("1010001001100000")
-    SN.append("0000100000001000")
-    HH.append("1010101010001010")
-
-    BD.append("1010000101110001")
-    SN.append("0000100000001000")
-    HH.append("1010101010001010")
-
-    # End external patterns
-
-    BD.append("1000100010001000")
-    SN.append("0000101001001000")
-    HH.append("0101010101010101")
-
-    BD.append("1100000001010000")
-    SN.append("0000101000001000")
-    HH.append("0101010101010101")
-
-    BD.append("1100000001010000")
-    SN.append("0000101000001000")
-    HH.append("1111111111111111")
-
-    BD.append("1001001001000100")
-    SN.append("0001000000010000")
-    HH.append("0101110010011110")
-
-    BD.append("1001001001000100")
-    SN.append("0001000000010000")
-    HH.append("1111111111111111")
-
-    # Be warned patterns < 16 steps can sound disjointed when using CV to select the pattern!
-
-    BD.append("10010000010010")
-    SN.append("00010010000010")
-    HH.append("11100110111011")
-
-    BD.append("1001000001001")
-    SN.append("0001001000001")
-    HH.append("1110011011101")
-
-    BD.append("100100000100")
-    SN.append("000100100000")
-    HH.append("111001101110")
-
-    BD.append("10010000010")
-    SN.append("00010010000")
-    HH.append("11100110111")
-
-    BD.append("10010000010")
-    SN.append("00010010000")
-    HH.append("11111010011")
-
-    BD.append("1001000010")
-    SN.append("0001000000")
-    HH.append("1111101101")
-
-    BD.append("100100010")
-    SN.append("000100000")
-    HH.append("111110111")
-
-    BD.append("10010010")
-    SN.append("00010000")
-    HH.append("11111111")
-
-    BD.append("1001001")
-    SN.append("0001000")
-    HH.append("1111111")
-
-    BD.append("100100")
-    SN.append("000100")
-    HH.append("111111")
-
-    BD.append("10000")
-    SN.append("00001")
-    HH.append("11110")
-
-    BD.append("1000")
-    SN.append("0000")
-    HH.append("1111")
-
-    BD.append("100")
-    SN.append("000")
-    HH.append("111")
